requirements.txt
requests==2.32.5
pandas==3.0.0
python-dotenv==1.2.1
numpy==2.4.1
matplotlib==3.10.8